# bpo-42808: Add PyType_Type.tp_vectorcall for type(obj) performance