--- a/Misc/NEWS.d/next/Core and Builtins/2021-01-02-05-10-58.bpo-42808.AOxgxl.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2021-01-02-05-10-58.bpo-42808.AOxgxl.rst
@@ -1,0 +1,1 @@
+Use the `Vectorcall` C calling convention for `type(obj)` calls.