--- a/Misc/NEWS.d/next/Core and Builtins/2021-01-02-05-10-58.bpo-42808.AOxgxl.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2021-01-02-05-10-58.bpo-42808.AOxgxl.rst
@@ -1,1 +1,2 @@
-Use the `Vectorcall` C calling convention for `type(obj)` calls.
+Simple calls to ``type(object)`` are now faster due to the
+``vectorcall`` calling convention. Patch by Dennis Sweeney.