@markshannon 2021-02-03
Yes, I do. `type(x)` is quite common in Python code (unlike the 3 argument form).