@sweeneyde 2021-02-05
Any update on this? I'm seeing some nice benchmarks:

```
.\python.bat -m pyperf timeit -s "import collections, itertools" "collections.deque(map(type, itertools.repeat(17, 1_000_000)), maxlen=0)"
Before: Mean +- std dev: 26.0 ms +- 0.8 ms
After:  Mean +- std dev: 11.6 ms +- 0.1 ms

.\python.bat -m pyperf timeit "type(17)"
Before: 51.6 ns +- 3.5 ns
After:  33.8 ns +- 2.1 ns
```